Prevent double rendering of head elements
